add input checking

+++ pkg/persist/interface.go
@@ -4,5 +4,5 @@ package persist
 type Db interface {
 	SetNX(string, interface{}, int64) error
 	Incr(string) (int64, error)
-	Reset(string, int) error
+	Reset(string, int64) error
 }
